Refactor AuthCallback and useAuth to implement exchangeAuthCodeForToken function; streamline authentication flow
